feature : folder 페이지에서 FolderMenu 컴포넌트 UI 분리
