+++ src/pages/folderPage/FolderPage.module.css
@@ -5,32 +5,6 @@
 .folderMenu {
 }
 
-.upperMenu {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-}
-
-.folderSelector {
-  padding: 0;
-}
-
-.addFolder {
-  display: none;
-}
-
-.title {
-}
-
-.editor {
-}
-
-.folderOption {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-}
-
 @media (min-width: 768px) {
   .linkList {
     display: grid;
